refactor: extract reconciliation orchestration
